fix: improve chat OAuth popup communication and UX
- Replace interval-based polling with postMessage/storage events for OAuth results
- Add immediate popup result delivery via postMessage
- Improve error handling with specific error messages
- Add server verification before marking authenticated
- Simplify auth form by removing loading state tracking
- Add token expiration validation in chat OAuth
- Fix popup management to prevent multiple windows

These changes improve the reliability and user experience of the OAuth
flow in the chat demo application without modifying the core OAuth
provider implementation.
